refactor: remove unnecessary component files
