+++ apps/tangle-website/src/components/index.ts
@@ -4,13 +4,10 @@ export * from './Heading';
 export * from './InternalOrExternalLink';
 export * from './LinkButton';
 export * from './Navbar';
-export * from './ParticipationTabTrigger';
-export * from './RoadMapItem';
 export * from './SectionDescription';
 export * from './SectionHeader';
 export * from './SectionTitle';
 export * from './SectionTitle2';
-export * from './TangleFeatureCard';
 // Section components
 export * from './sections';
 // Svg components
