+++ apps/tangle-website/src/components/sections/FeaturesSection.tsx
@@ -1,3 +1,6 @@
+import { FC } from 'react';
+import Image from 'next/image';
+import { Typography } from '@webb-tools/webb-ui-components';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { FreeMode } from 'swiper';
 import 'swiper/css';
@@ -7,9 +10,17 @@ import {
   SectionHeader,
   SectionTitle,
   SectionDescription,
-  TangleFeatureCard,
+  LinkButton
 } from '..';
 import { TANGLE_OVERVIEW_URL, WANT_TO_LEARN_MORE_URL } from '../../constants';
+
+interface TangleFeatureCardProps {
+  img: string;
+  index: number;
+  title: string;
+  description: string;
+  link: string;
+}
 
 const tangleFeatures = [
   {
@@ -89,5 +100,31 @@ export const FeaturesSection = () => {
         })}
       </Swiper>
     </section>
+  );
+};
+
+const TangleFeatureCard: FC<TangleFeatureCardProps> = (props) => {
+  const { img, index, title, description, link } = props;
+  return (
+    <div className="bg-mono-0 rounded-lg overflow-hidden flex flex-col md:w-[300px] min-h-[647px] md:min-h-min md:h-full lg:h-auto flex-1">
+      <div className="relative h-[150px] w-full object-contain">
+        <Image src={img} alt={title} fill />
+      </div>
+      <div className="py-[42px] px-6 flex flex-col justify-between flex-1">
+        <div>
+          <p className="mono1 mb-4">0{index}</p>
+          <hr />
+          <Typography variant="h4" fw="bold" className="mt-4 mb-6">
+            {title}
+          </Typography>
+          <SectionDescription className="text-left !text-xl !leading-8 md:!text-lg md:!leading-[32.4px] lg:!text-xl lg:!leading-8">
+            {description}
+          </SectionDescription>
+        </div>
+        <LinkButton href={link} className="mt-4">
+          Learn More
+        </LinkButton>
+      </div>
+    </div>
   );
 };
