+++ apps/tangle-website/src/components/sections/ParticipationMechanicsSection.tsx
@@ -1,8 +1,4 @@
 import { useState } from 'react';
-import { Swiper, SwiperSlide } from 'swiper/react';
-import { FreeMode } from 'swiper';
-import 'swiper/css';
-import 'swiper/css/free-mode';
 import {
   TabContent,
   TabsList,
@@ -10,28 +6,23 @@ import {
   TabTrigger,
   Typography,
 } from '@webb-tools/webb-ui-components';
+import { Swiper, SwiperSlide } from 'swiper/react';
+import { FreeMode } from 'swiper';
+import 'swiper/css';
+import 'swiper/css/free-mode';
 
-import {
-  SectionDescription,
-  SectionHeader,
-  SectionTitle,
-  ParticipationTabTrigger,
-} from '..';
-
+import { SectionDescription, SectionHeader, SectionTitle } from '..';
 import {
   WEBB_DOCS_URL,
   NODE_OPERATORS_URL,
   POLKADOT_TANGLE_URL,
 } from '../../constants';
 
+interface TangleFeatureCardProps {
+  tabName: string;
+}
+
 const tabsContent = {
-  // Crowdloan: {
-  //   title: 'Participate Tangle’s Crowdloan',
-  //   description:
-  //     'A community-backed launch that allows people to contribute by locking up their KSM temporarily. Tangle’s native token (TNT) will be distributed based on the amount of KSM locked. Join waitlist, and be sure you’re eligible to participate with tokens on hand.',
-  //   linkText: 'Contribute Now',
-  //   linkUrl: WEBB_DOCS_URL,
-  // },
   'Claim Airdrop': {
     title: 'Claim Airdrop',
     description:
@@ -79,6 +70,7 @@ export const ParticipationMechanicsSection = () => {
           value={activeTab}
           onValueChange={(nextTab) => setActiveTab(nextTab as TabTypes)}
         >
+
           {/* Desktop + Mobile Tab Triggers */}
           <TabsList aria-label="tabs" className="lg:mx-0">
             <div className="grid md:hidden lg:grid w-full grid-cols-2 lg:grid-cols-4 gap-3 mb-6">
@@ -144,5 +136,24 @@ export const ParticipationMechanicsSection = () => {
         </TabsRoot>
       </div>
     </section>
+  );
+};
+
+const ParticipationTabTrigger: React.FC<TangleFeatureCardProps> = (props) => {
+  const { tabName } = props;
+  return (
+    <>
+      <div className="participation-tab w-full aspect-square flex justify-center items-center rounded-lg">
+        <div className="w-full text-inherit">
+          <div className="flex flex-col items-center gap-2 text-inherit">
+            <div className="w-12 h-12 bg-mono-40 rounded-full" />
+            <p className="text-[16px] leading-[25.6px] md:text-[24px] md:leading-[40px] font-bold text-inherit">
+              {tabName}
+            </p>
+          </div>
+        </div>
+      </div>
+      <div className="participation-tab-polygon w-0 h-0 border-transparent border-solid border-x-[8px] border-t-[16px]" />
+    </>
   );
 };
